batch: Move merge candidate values
Batch merge owns structural merge execution plus the value objects used to
describe reviewed merge candidates. Operation candidate planning imports
those objects through the merge implementation module.

The project lacks a focused boundary for candidate data. That makes callers
depend on the merge engine when they only need resolution shapes, which
expands the import surface around one of the largest batch modules.

This commit begins separating candidate preview planning from merge
execution by moving the candidate dataclasses into batch.merge_candidates,
importing them privately inside merge.py, and pointing
operation_candidates.py at the smaller data module.

Subsequent commits will add tests that pin the candidate boundary and
prevent batch.merge from becoming the value-object import path again.
